feat: refactor bookmark search page (#739)
* feat: refactor bookmark search page
