+++ packages/webapp/pages/bookmarks.tsx
@@ -4,8 +4,8 @@ import { NextSeo } from 'next-seo';
 import {
   getBookmarkFeedLayout,
   bookmarkFeedLayoutProps,
-} from '../../components/layouts/BookmarkFeedPage';
-import { defaultOpenGraph, defaultSeo } from '../../next-seo';
+} from '../components/layouts/BookmarkFeedPage';
+import { defaultOpenGraph, defaultSeo } from '../next-seo';
 
 const seo: NextSeoProps = {
   title: `Your daily.dev bookmarks`,
